Refactor Operations/AggregatorOperations for convenience in preparing to continue removing *args
