+++ lib/cascading.rb
@@ -14,6 +14,7 @@ require 'cascading/expr_stub'
 require 'cascading/flow'
 require 'cascading/mode'
 require 'cascading/operations'
+require 'cascading/aggregator_operations'
 require 'cascading/identity_operations'
 require 'cascading/filter_operations'
 require 'cascading/regex_operations'
